# Reduce execution time of stream_unfetched_token_instances query