## Motivation

The query https://github.com/poanetwork/blockscout/blob/da54650a4a559393d29a499a486b51d1477789ef/apps/explorer/lib/explorer/chain.ex#L2969 executes more than 1 hour on ETH Mainnet instance

This PR aims to increase the performance of this query.

## Changelog

Decrease the subset before joining the tables

Before the changes query looked like this:

```
SELECT DISTINCT ON (t0."token_contract_address_hash", t0."token_id") t0."token_contract_address_hash", t0."token_id" FROM "token_transfers" AS t0
INNER JOIN "tokens" AS t1 ON t1."contract_address_hash" = t0."token_contract_address_hash"
LEFT OUTER JOIN "token_instances" AS t2 ON (t0."token_id" = t2."token_id")
AND (t0."token_contract_address_hash" = t2."token_contract_address_hash")
WHERE (((t1."type" = 'ERC-721')
AND t2."token_id" IS NULL) AND NOT (t0."token_id" IS NULL))
LIMIT 1;
```

EXPLAIN ANALYZE:

```
Planning Time: 0.520 ms
Execution Time: 0.851 ms
```

After the changes:

```
SELECT DISTINCT q."token_contract_address_hash", q."token_id" FROM (SELECT
  t0."token_contract_address_hash", t0."token_id" FROM "token_transfers" AS t0
INNER JOIN (SELECT * FROM "tokens" WHERE tokens."type" = 'ERC-721') AS t1 ON t1."contract_address_hash" = t0."token_contract_address_hash"
LEFT OUTER JOIN "token_instances" AS t2 ON (t0."token_id" = t2."token_id")
AND (t0."token_contract_address_hash" = t2."token_contract_address_hash")
AND t2."token_id" IS NULL AND NOT (t0."token_id" IS NULL)
) q;
```

EXPLAIN ANALYZE:

```
Planning Time: 0.423 ms
Execution Time: 0.592 ms
```

## Checklist for your Pull Request (PR)

  - [x] I added an entry to `CHANGELOG.md` with this PR
  - [x] If I added new functionality, I added tests covering it.
  - [x] If I fixed a bug, I added a regression test to prevent the bug from silently reappearing again.
  - [x] I checked whether I should update the docs and did so by submitting a PR to https://github.com/blockscout/docs
  - [x] If I added/changed/removed ENV var, I submitted a PR to https://github.com/blockscout/docs to update the list of env vars at https://github.com/blockscout/docs/blob/master/for-developers/information-and-settings/env-variables.md and I updated the version to `master` in the Version column. Changes will be reflected in this table: https://docs.blockscout.com/for-developers/information-and-settings/env-variables. 
  - [x] If I add new indices into DB, I checked, that they are not redundant with PGHero or other tools