@vbaranov 2019-12-10
@ayrat555 it is again `test ethsupply returns total supply from DB`. It is failed through all branches except master.