@vbaranov 2019-12-10
@ayrat555 thanks, column name was wrong in the `distinct` select from the subquery. Fixed https://github.com/poanetwork/blockscout/pull/2914/commits/a7d9f0fcf2235bb27e666d1c6a6ea1df596f1f59